Notes_№3/notesRest/requirements.txt
accessify==0.3.1
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.6.3
asgiref==3.8.1
asttokens==2.2.1
async-timeout==4.0.3
asyncio==3.4.3
atomicwrites==1.4.1
attrs==22.1.0
autoflake==2.0.1
autoimport==1.3.3
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
cheap-repr==0.5.1
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
Django==3.2.3
django-cors-headers==4.3.1
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dnspython==2.4.2
drf-yasg==1.21.7
email-validator==2.1.0.post1
exceptiongroup==1.1.0
executing==1.2.0
Faker==12.0.1
fastapi==0.105.0
flake8==6.0.0
flake8-docstrings==1.6.0
flake8-isort==6.0.0
flatbuffers==23.5.26
frozenlist==1.4.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
humanfriendly==10.0
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
isort==5.12.0
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
maison==1.4.0
MarkupSafe==2.1.2
mccabe==0.7.0
mixer==7.1.2
mpmath==1.3.0
multidict==6.0.4
npm==0.1.1
numpy==1.26.2
oauthlib==3.2.2
onnxruntime==1.16.2
opencv-python==4.8.1.78
optional-django==0.1.0
orjson==3.9.10
packaging==21.3
pandas==2.2.2
Pillow==9.3.0
pluggy==0.13.1
protobuf==4.25.1
py==1.11.0
pyaes==1.6.1
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.5.2
pydantic-extra-types==2.2.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
pydocstyle==6.3.0
pyflakes==3.0.1
Pygments==2.15.0
PyJWT==2.1.0
pyparsing==3.0.9
pyprojroot==0.2.0
pyreadline3==3.4.1
Pyrogram==2.0.97
PySocks==1.7.1
pyTelegramBotAPI==4.14.0
pytest==6.2.4
pytest-django==4.5.2
pytest-ini==0.33
pytest-pythonpath==0.7.3
pytest-runner==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
python-telegram-bot==13.7
python3-openid==3.2.0
pytils==0.3
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
requests==2.26.0
requests-oauthlib==1.3.1
sh==2.0.1
six==1.16.0
sniffio==1.3.0
snoop==0.4.3
snowballstemmer==2.2.0
social-auth-app-django==4.0.0
social-auth-core==4.4.0
sorl-thumbnail==12.7.0
SQLAlchemy==2.0.30
sqlparse==0.4.3
starlette==0.27.0
sympy==1.12
telebot==0.0.5
toml==0.10.2
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2022.7
tzlocal==4.2
ujson==5.9.0
unicorn==2.0.1.post1
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.25.0
varname==0.8.1
vk-api==11.9.9
vk-captcha==2.0
watchfiles==0.21.0
webcolors==1.13
websockets==12.0
xdg==5.1.1
yarl==1.9.2
